refactor(e2e): fold regression tests into GlossaryTermRelationSettings.spec.ts
Removes the standalone GlossaryTermRelationSettingsRegression.spec.ts and
inlines the three parallel-execution regression tests at the end of the
existing spec, keeping all relation-settings coverage co-located in one file.
Updates impact-map.json to drop the now-unnecessary separate spec reference.
